requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Streamlit UI
streamlit>=1.30.0

# FastAPI Backend
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Document Parsing
pypdf2>=3.0.0
python-docx>=1.1.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP and API
requests>=2.31.0
httpx>=0.26.0

# Environment Management
python-dotenv>=1.0.0

# Caching (optional)
redis>=5.0.0

# Utilities
python-dateutil>=2.8.2
tiktoken>=0.5.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
